changed hostel lat and long to double instead of float Allows for the correct number of decimal places I need

+++ migrations/20190124135853_hostels.js
@@ -2,8 +2,8 @@ exports.up = function (knex, Promise) {
     return knex.schema.createTable('hostels', (hostel) => {
         hostel.increments('id')
         hostel.string('name')
-        hostel.float('latitude')
-        hostel.float('longitude')
+        hostel.double('latitude')
+        hostel.double('longitude')
         hostel.string('image')
     })
 }
